fix: JourneyVector namespaced
