--- app/graphql/types/journey_vector_type.rb
@@ -1,8 +1,0 @@
-module Types
-  class JourneyVectorType < Types::BaseObject
-    field :from, String, null: true
-    field :to, String, null: true
-    field :uri, String, null: true
-    field :via, String, null: true
-  end
-end
