+++ app/graphql/types/tfl/entities/journey_planner/itinerary_result_type.rb
@@ -7,7 +7,7 @@ module Types
           field :recommendedMaxAgeMinutes, Integer, null: true
           field :lines, [Types::Tfl::Entities::LineType, null: true], null: true
           field :journeys, [Types::JourneyType, null: true], null: true
-          field :journeyVector, Types::JourneyVectorType, null: true
+          field :journeyVector, Types::Tfl::Entities::JourneyPlanner::JourneyVectorType, null: true
           field :searchCriteria, Types::SearchCriteriaType, null: true
           field :cycleHireDockingStationData, Types::CycleHireDockingStationDataType, null: true
         end
