+++ app/graphql/types/tfl/entities/journey_planner/journey_vector_type.rb
@@ -1,0 +1,14 @@
+module Types
+  module Tfl
+    module Entities
+      module JourneyPlanner
+        class JourneyVectorType < Types::BaseObject
+          field :from, String, null: true
+          field :to, String, null: true
+          field :uri, String, null: true
+          field :via, String, null: true
+        end
+      end
+    end
+  end
+end
